Refactored the entities to parse only the necessary information.
